pyproject.toml
[project]
name = "rdp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aim>=3.29.1",
    "einops>=0.8.2",
    "hydra-core>=1.3.2",
    "lightning>=2.6.1",
    "matplotlib>=3.10.8",
    "numba>=0.64.0",
    "omegaconf>=2.3.0",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "pytorch-lightning>=2.6.1",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "transformers>=5.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/RDP"]
